requirements.txt
cryptography==43.0.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
mysql-connector-python==9.0.0
pandas==2.2.2
PyMySQL==1.1.1
python-dotenv==1.0.1
usports-basketball==2.0.1
